refactor: Celery 워커 진입점을 app.worker로 분리 (웹 계층 미의존)
- create_celery + worker_shutting_down 훅을 app/worker/app.py로 이동,
  celery_utils.py는 re-export shim 유지 (Phase 2에서 제거)
- 워커 Dockerfile 5종 CMD: app.main.celery → app.worker.app
  (워커가 더 이상 마이그레이션·웹 앱 조립을 실행하지 않음)
- 태스크 이름 불변: plugin_task:build_plugin_task,
  workflow_task:process_data_task (name= 기존 명시 확인)
- import-linter no-web-in-worker 계약 활성화 (1 kept, 0 broken)

배포 주의: 구워커 큐 드레인 후 웹(마이그레이션) → 워커 순 기동
