+++ backend/.importlinter
@@ -6,25 +6,25 @@
 # Purpose: define the architectural import contracts up front (during PR-1,
 # the safety-net phase) so they can be switched on as the refactoring lands.
 #
-# STATUS: the `no-web-in-worker` contract is defined but INACTIVE. There is no
-# `app.worker` package yet — the Celery worker still boots via `app.main`
-# (Dockerfile.celery: `celery -A app.main.celery worker`). This contract becomes
-# meaningful in PR-2 (phase-1-app-assembly), where `app.worker` is introduced and
-# the worker entry point stops importing the web layer. To ACTIVATE in PR-2:
-#   1. Uncomment the [importlinter:contract:no-web-in-worker] block below.
-#   2. Ensure `app.worker` exists.
-#   3. Add `import-linter` to the backend dev/test dependencies and wire
-#      `lint-imports` into CI.
+# STATUS: ACTIVATED in PR-2 (phase-1-app-assembly). `app.worker` now exists and the
+# Celery worker boots via `celery -A app.worker.app worker` — it no longer imports
+# the web layer (`app.main`). See the contract below.
+#
+# NOTE on scope: only `app.main` is forbidden. The web router package `app.routes`
+# cannot be blanket-forbidden yet because the Celery tasks module still lives at
+# `app.routes.celery_tasks` (the worker legitimately includes it). The
+# `app.routes.api` web-assembly module is forbidden as well since the task-chain
+# does not depend on it. Tightening `app.routes` fully happens in a later PR that
+# relocates the tasks module out of the web-router package.
 
 [importlinter]
 root_package = app
 
-# --- Activated in PR-2 (phase-1-app-assembly) — keep commented until app.worker exists ---
-# [importlinter:contract:no-web-in-worker]
-# name = worker must not depend on the web layer
-# type = forbidden
-# source_modules =
-#     app.worker
-# forbidden_modules =
-#     app.main
-#     app.api
+[importlinter:contract:no-web-in-worker]
+name = worker must not depend on the web layer
+type = forbidden
+source_modules =
+    app.worker
+forbidden_modules =
+    app.main
+    app.routes.api
